requirements.txt
Django==1.6.2
South==0.8.4
Whoosh==2.6.0
argparse==1.2.1
beautifulsoup4==4.3.2
coverage==3.7.1
django-ajax-selects==1.3.4
django-coverage==1.2.4
django-haystack==2.1.0
django-nose==1.2
django-registration==1.0
lxml==3.3.3
nose==1.3.3
psycopg2==2.5.2
requests==2.2.1
selenium==2.41.0
wsgiref==0.1.2
